Move ExprOp fixpoint wrappers to their own file.
Also, fix bad indentation.
