+++ core/src/main/scala/slamdata/engine/physical/mongodb/reshape.scala
@@ -25,7 +25,7 @@ import slamdata.engine.{Error, RenderTree, Terminal, NonTerminal, RenderedTree}
 import slamdata.engine.fp._
 import slamdata.engine.javascript._
 import slamdata.engine.physical.mongodb.accumulator._
-import slamdata.engine.physical.mongodb.expression._; import DSL._
+import slamdata.engine.physical.mongodb.expression._
 
 final case class Grouped(value: ListMap[BsonField.Leaf, Accumulator]) {
   def bson = Bson.Doc(value.map(t => t._1.asText -> groupBson(t._2)))
