+++ core/src/test/scala/slamdata/engine/physical/mongodb/optimize.scala
@@ -10,7 +10,7 @@ import slamdata.engine.{TreeMatchers}
 
 class OptimizeSpecs extends Specification with TreeMatchers {
   import slamdata.engine.physical.mongodb.accumulator._
-  import slamdata.engine.physical.mongodb.expression._; import DSL._
+  import slamdata.engine.physical.mongodb.expression._
   import Workflow._
   import IdHandling._
