+++ core/src/test/scala/slamdata/engine/physical/mongodb/pipeline.scala
@@ -13,7 +13,7 @@ import slamdata.specs2._
 
 class PipelineSpec extends Specification with ScalaCheck with DisjunctionMatchers with ArbBsonField with PendingWithAccurateCoverage {
   import slamdata.engine.physical.mongodb.accumulator._
-  import slamdata.engine.physical.mongodb.expression._; import DSL._
+  import slamdata.engine.physical.mongodb.expression._
   import Workflow._
   import ArbitraryExprOp._
